CROWN SUMMONING FIXES | Makes you no longer need to wear any hat to use throne basic commands. (#180)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Zethcode.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Zethcode being fixed is always good. Right now, if you are not wearing
something on your head slot, you cannot use the throne regardless, even
Summon Crown.

This removes that, and changes how the Crown works.

Should the Crown not be present on a living human, it may immediately be
summoned back to the keep (like before), though with a visible message
instead.

This means that unlike before if the King was decapitated and the crown
hidden, or thrown into the acid pit, there is now an actual way to
retrieve the crown.

![image](https://github.com/Rotwood-Vale/Rotwood-Vale/assets/65188584/5a9455d2-8212-4289-bf7b-a87a42ce3436)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Azarak <azarak10@gmail.com>