Gamemode changes (#275)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

It seems the consensus is that we need more spice, so here's some
changes

The major gamemode now has a chance to be:
35% rebellion (previously 33%)
45% vamps & werewolves (previously 33%)
20% extended  (previously 33%)

The minor gamemodes now have this chance to add itself (can add
multiple, and could before too):
45% bandits (previously ~35%)
45% aspirant (previously ~35%)
10% maniac (previously ~8,7%)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->