Leper Flaw
Adds Leprosy as a flaw. Currently doesn't do much other than sometimes cause limbs to spasm and gives notifications regarding sickness. As it stands this is mostly a roleplay Flaw as it doesn't really have the main symptoms and flaws the actual one would have, but I still made it for people to build off of.