No More Sword RNG for Paladin!
I find it really, really godawful that Paladin literally gets stuck with only a regular Steel Sword 70% of the time, and no shield or anything to go with it. Steel Swords aren't that good even when two-handed. This fixes it so at least they get a guaranteed Bastard Sword when they spawn in. I mean, they're basically diet Templars, they don't even get as good armor or gear as them or the same stats... at least let 'em have a good sword. Sheesh.