Update sewing.dm
Cloth glove crafting renamed to "fingerless cloth gloves", which is what the craft actually makes.
Headscarves can now be untied to make cloth, just like the headband. (Headband is still the fastest to use for skill grinding as it's first on the crafting list.)
Sewing crafting list has been re-arranged to make it easier to make the armored items and to find related items.